fix(graph): the sibling I missed, and two retention paths that stopped short
Three gaps the stop-time review found in the previous reconciliation. Two of
them are the same mistake in different places: fixing the instance and leaving
the sibling.

RANKING, the sibling. I fixed /people/{id}/network to over-fetch, rank by
warmth and then cap — and left the who_knows AGENT tool capping by last
contact. The answer to "who knows them best" therefore depended on whether a
human or an agent asked. Both paths now take the same three steps.

RANKING, deal coverage. our_side came back in last-contact order, which is not
what a coverage view asks and is not deterministic across loads. Ranked with
the id tie-break every ordered payload here carries.

RETENTION, the sibling. `activity/archive` invalidated the interaction edges;
`activity/erase` did not, though it archives the activity too. An erased
conversation kept counting toward a relationship score until the nightly
rebuild.

RETENTION, the address arm. The anonymize sweep matched participants and ghosts
on person_id ALONE, so the subject's raw email address survived — and the
address arm of a participant row exists precisely for someone with no person
id. It now matches both, reading the addresses BEFORE person_email is deleted:
the same ordering trap the eraser hit with the subject's name, one column over.

Splitting the AI-store sweeps out of retention.go made three tables look
unswept, because the PII census was keyed to a single filename. It is a list
now — a census that reports a refactor as a compliance regression trains people
to ignore it.
